fix: babysit round 1, 8 findings + Pattern A enhancement
Pre-PR review feedback addressed across 12 reviewer agents.

Lint correctness:

- Factory ghost classification now requires registered default-disabled gating setting (CR finding); a missing flag is out of scope rather than implicitly default-disabled

- _find_factory_function return type widened to FunctionDef | AsyncFunctionDef (Gemini medium); callers updated

- Help text drops the nonexistent --paths advertisement (CR finding)

- Scope-aware ghost detection: start-gate matching now verifies the gate variable's binding annotation matches the ghost class, defending against same-name collisions in unrelated scopes (CR major finding)

Module split (CR major + 800-line ceiling per CLAUDE.md):

- scripts/check_setting_to_startup_trace.py shrinks from 1550 lines to 259-line CLI shim

- Logic moves to four sibling modules following the scripts/check_web_design_system pattern (sibling-import dance for standalone vs package invocation):

  * _setting_to_startup_trace_models.py (166 lines): dataclasses, Literal types, suppression marker

  * _setting_to_startup_trace_loader.py (205 lines): settings inventory

  * _setting_to_startup_trace_ghosts.py (661 lines): lifecycle parsing + ghost detection + scope-aware matching

  * _setting_to_startup_trace_violations.py (442 lines): Pattern A + matchers + scan_repo + baseline

Test split:

- tests/unit/scripts/test_check_setting_to_startup_trace.py shrinks from 1285 lines to 217-line smoke + inventory file

- Helpers extracted to tests/unit/scripts/_setting_to_startup_trace_helpers.py

- Per-category test files added: ghosts (512 lines), baseline (364 lines), pattern_a (160 lines)

Documentation:

- Bootstrap-wiring trace subsection in docs/reference/configuration-precedence.md now describes all three matchers (gating-namespace, class-file containment, Pattern A direct ConfigResolver) plus debugging guidance for cross-namespace flags (CR finding)

- Audit-run callout dropped (forensic narrative)

Skipped (factually wrong):

- 5x gemini-code-assist 'PEP 758 except syntax' findings: false positive. Python 3.14 PEP 758 (accepted, https://peps.python.org/pep-0758/) explicitly allows unparenthesized except A, B: when not binding. CLAUDE.md mandates this form; ruff enforces it on 3.14. Tests pass on 3.14.4. An explanatory comment was added near the top of the script

- 1x gemini-code-assist '_RESOLVER_MIN_ARGS docstring misplaced': false positive. Both _RESOLVER_GET_METHODS and _RESOLVER_MIN_ARGS have correctly-attached docstrings

Total: 36 unit tests pass across 4 split files; lint exits 0 against real repo
